[VAS-821] feat: CI's station detail page new design
