+++ src/pages/stations/detail/__tests__/StationDetailPage.test.tsx
@@ -84,7 +84,6 @@ describe('<StationDetailPage />', () => {
     );
 
     expect(screen.getByText('stationDetailPage.associates')).toBeInTheDocument();
-    expect(screen.queryByTestId('station-detail-op')).not.toBeInTheDocument();
   });
 
   test('Test Render station detail with role operator', async () => {
