+++ src/pages/stations/detail/components/DetailButtonsStation.tsx
@@ -1,19 +1,168 @@
-/* eslint-disable sonarjs/no-identical-functions */
-import {Button, Stack} from '@mui/material';
-import {generatePath, Link} from 'react-router-dom';
-import {useTranslation} from 'react-i18next';
+import { useMemo, useState } from 'react';
+import { Box, Button, Stack, TextField, Typography } from '@mui/material';
+import { generatePath, Link } from 'react-router-dom';
+import { useErrorDispatcher, useLoading } from '@pagopa/selfcare-common-frontend';
+import {
+  ContentCopy,
+  Delete,
+  ManageAccounts,
+  MiscellaneousServices,
+  ModeEdit,
+} from '@mui/icons-material';
+import { useTranslation } from 'react-i18next';
 import ROUTES from '../../../../routes';
-import {StationFormAction} from '../../../../model/Station';
-import {WrapperStatusEnum} from '../../../../api/generated/portal/StationDetailResource';
-import {useUserRole} from "../../../../hooks/useUserRole";
+import { StationFormAction } from '../../../../model/Station';
+import {
+  StationDetailResource,
+  WrapperStatusEnum,
+} from '../../../../api/generated/portal/StationDetailResource';
+import { useUserRole } from '../../../../hooks/useUserRole';
+import { updateWrapperStationWithOperatorReview } from '../../../../services/stationService';
+import { LOADING_TASK_STATION_DETAILS_REQUEST_EDIT } from '../../../../utils/constants';
+import GenericModal from '../../../../components/Form/GenericModal';
 
-type Props = {
-  status: WrapperStatusEnum | undefined;
-  stationCode: string;
+const ModalContent = ({
+  setShowModal,
+  stationDetail,
+  setStationDetail,
+}: {
+  setShowModal: (value: boolean) => void;
+} & Props) => {
+  const { t } = useTranslation();
+
+  const addError = useErrorDispatcher();
+  const setLoading = useLoading(LOADING_TASK_STATION_DETAILS_REQUEST_EDIT);
+
+  const [input, setInput] = useState<string>(stationDetail?.note ?? '');
+
+  const sendEditRequest = () => {
+    setLoading(true);
+    updateWrapperStationWithOperatorReview({
+      stationCode: stationDetail?.stationCode ?? '',
+      ciTaxCode: stationDetail?.brokerCode ?? '',
+      note: input,
+    })
+      .then((data: StationDetailResource) => {
+        setStationDetail(data);
+      })
+      .catch((reason) =>
+        addError({
+          id: 'PUT_STATION_DETAILS_REQUEST_EDIT',
+          blocking: false,
+          error: reason as Error,
+          techDescription: `An error occurred while getting updating the station with operator's note`,
+          toNotify: true,
+          displayableTitle: t('general.errorTitle'),
+          displayableDescription: t('stationDetailPageValidation.modal.error'),
+          component: 'Toast',
+        })
+      )
+      .finally(() => {
+        setShowModal(false);
+        setLoading(false);
+      });
+  };
+  return (
+    <>
+      <Typography variant="h6">{t('stationDetailPageValidation.modal.title')}</Typography>
+      <Typography variant="body1" sx={{ my: 2 }}>
+        {t('stationDetailPageValidation.modal.subtitle')}
+      </Typography>
+      <TextField
+        fullWidth
+        id="requestInput"
+        name="requestInput"
+        required
+        multiline
+        placeholder={t('stationDetailPageValidation.modal.placeholder')}
+        size="small"
+        value={input}
+        onChange={(e) => setInput(e.target.value)}
+        helperText={t('stationDetailPageValidation.modal.helperText')}
+        inputProps={{
+          'data-testid': 'requestInput',
+          maxLength: 200,
+        }}
+      />
+      <Box display="flex" justifyContent={'flex-end'} mt={2}>
+        <Button
+          variant="outlined"
+          sx={{ mr: 1 }}
+          onClick={() => setShowModal(false)}
+          data-testid="cancel-button-test"
+        >
+          {t('general.turnBack')}
+        </Button>
+        <Button
+          variant="contained"
+          onClick={() => sendEditRequest()}
+          data-testid="confirm-button-test"
+        >
+          {t('general.confirmAndSend')}
+        </Button>
+      </Box>
+    </>
+  );
 };
 
-const DetailButtonsStation = ({ status, stationCode }: Props) => {
+type Props = {
+  stationDetail: StationDetailResource;
+  setStationDetail: (value: any) => void;
+};
+
+const DetailButtonsStation = ({ stationDetail, setStationDetail }: Props) => {
   const { t } = useTranslation();
+  const { userIsPagopaOperator } = useUserRole();
+
+  const status = stationDetail?.wrapperStatus;
+  const stationCode = stationDetail?.stationCode;
+
+  const [showModal, setShowModal] = useState<boolean>(false);
+
+  const editPath = () =>
+    generatePath(ROUTES.STATION_EDIT, {
+      stationId: stationCode,
+      actionId: StationFormAction.Edit,
+    });
+
+  if (userIsPagopaOperator) {
+    return (
+      <>
+        <Stack spacing={2} direction="row" flexWrap={'wrap'} justifyContent={'flex-end'}>
+          {(status === WrapperStatusEnum.TO_CHECK ||
+            status === WrapperStatusEnum.TO_CHECK_UPDATE) && (
+            <Button variant="outlined" onClick={() => setShowModal(true)}>
+              {t('stationDetailPage.stationOptions.requestEdit')}
+            </Button>
+          )}
+          <Button
+            component={Link}
+            to={editPath}
+            variant="contained"
+            data-testid="edit-btn-ope-sts-approved"
+          >
+            {t(
+              stationDetail?.wrapperStatus !== WrapperStatusEnum.APPROVED
+                ? 'stationDetailPage.stationOptions.approveAndValidate'
+                : 'general.modify'
+            )}
+          </Button>
+        </Stack>
+        {showModal && (
+          <GenericModal
+            openModal={showModal}
+            renderContent={() => (
+              <ModalContent
+                setShowModal={setShowModal}
+                setStationDetail={setStationDetail}
+                stationDetail={stationDetail}
+              />
+            )}
+          />
+        )}
+      </>
+    );
+  }
 
   return (
     <Stack spacing={2} direction="row" flexWrap={'wrap'} justifyContent={'flex-end'}>
@@ -25,9 +174,19 @@ const DetailButtonsStation = ({ status, stationCode }: Props) => {
             color="error"
             variant="outlined"
             disabled={true}
+            startIcon={<Delete />}
             // onClick={() => ''}
           >
-            {t('stationDetailPage.stationOptions.deleteRequired')}
+            {t("general.delete")}
+          </Button>
+          <Button
+            component={Link}
+            to={generatePath(ROUTES.STATION_EC_LIST, { stationId: stationCode })}
+            color="primary"
+            startIcon={<MiscellaneousServices />}
+            variant="outlined"
+          >
+            {t('stationDetailPage.manageEC')}
           </Button>
           <Button
             component={Link}
@@ -37,6 +196,7 @@ const DetailButtonsStation = ({ status, stationCode }: Props) => {
                 actionId: StationFormAction.Duplicate,
               })
             }
+            startIcon={<ContentCopy />}
             variant="outlined"
             data-testid="duplicate-btn-sts-approved"
           >
@@ -44,47 +204,30 @@ const DetailButtonsStation = ({ status, stationCode }: Props) => {
           </Button>
           <Button
             component={Link}
-            to={() =>
-              generatePath(ROUTES.STATION_EDIT, {
-                stationId: stationCode,
-                actionId: StationFormAction.Edit,
-              })
-            }
+            to={editPath}
             variant="contained"
+            startIcon={<ModeEdit />}
             data-testid="edit-btn-sts-approved"
           >
             {t('stationDetailPage.stationOptions.editStation')}
-          </Button>
-        </>
-      ) : status === WrapperStatusEnum.TO_FIX ? (
-        <>
-          <Button
-            // component={Link}
-            // to={() => ''}
-            variant="contained"
-            // TBD
-            disabled={true}
-            data-testid="edit-btn-sts-fix"
-          >
-            {t('stationDetailPage.stationOptions.correctStation')}
           </Button>
         </>
       ) : (
-        <>
-          <Button
-            component={Link}
-            to={() =>
-              generatePath(ROUTES.STATION_EDIT, {
-                stationId: stationCode,
-                actionId: StationFormAction.Edit,
-              })
-            }
-            variant="contained"
-            data-testid="edit-btn-sts-approved"
-          >
-            {t('stationDetailPage.stationOptions.editStation')}
-          </Button>
-        </>
+        <Button
+          component={Link}
+          to={editPath}
+          variant="contained"
+          startIcon={<ModeEdit />}
+          data-testid="edit-btn-sts-approved"
+        >
+          {t(
+            `stationDetailPage.stationOptions.${
+              status === WrapperStatusEnum.TO_FIX || status === WrapperStatusEnum.TO_FIX_UPDATE
+                ? 'correctStation'
+                : 'editStation'
+            }`
+          )}
+        </Button>
       )}
     </Stack>
   );
