+++ src/pages/stations/detail/components/__tests__/StationDetails.test.tsx
@@ -1,241 +1,158 @@
 import React from 'react';
 import {ThemeProvider} from '@mui/system';
 import {theme} from '@pagopa/mui-italia';
-import {cleanup, fireEvent, render, screen} from '@testing-library/react';
+import {cleanup, fireEvent, render, screen, waitFor} from '@testing-library/react';
 import {MemoryRouter, Route} from 'react-router-dom';
 import {store} from '../../../../../redux/store';
 import {createMemoryHistory} from 'history';
 import {Provider} from 'react-redux';
-import StationDetails from '../StationDetails';
 import {
-    RedirectProtocolEnum,
-    StationDetailResource,
-    WrapperStatusEnum,
+  RedirectProtocolEnum,
+  StationDetailResource,
+  WrapperStatusEnum,
 } from '../../../../../api/generated/portal/StationDetailResource';
-import {partiesActions} from '../../../../../redux/slices/partiesSlice';
-import {ecAdminSignedDirect} from '../../../../../services/__mocks__/partyService';
-import {ROLE} from "../../../../../model/RolePermission";
+import {Protocol4ModEnum, ProtocolEnum} from '../../../../../api/generated/portal/StationDetailsDto';
+import StationDetails from '../StationDetails';
 import * as useUserRole from "../../../../../hooks/useUserRole";
-
-beforeEach(() => {
-    jest.spyOn(console, 'error').mockImplementation(() => {
-    });
-    jest.spyOn(console, 'warn').mockImplementation(() => {
-    });
-});
-
-jest.mock('../../../../components/commonFunctions');
-jest.mock("../../../../../hooks/useUserRole");
-
-afterEach(cleanup);
+import {ROLE} from "../../../../../model/RolePermission";
 
 const nodeCrypto = require('crypto');
 // @ts-ignore
 window.crypto = {
-    getRandomValues: function (buffer) {
-        return nodeCrypto.randomFillSync(buffer);
-    },
+  getRandomValues: function (buffer) {
+    return nodeCrypto.randomFillSync(buffer);
+  },
 };
 const genPassword = crypto.getRandomValues(new Uint32Array(1)).toString();
 
-export const mockedFullStationApproved: StationDetailResource = {
-    stationCode: '81001870922_06',
-    enabled: true,
-    brokerDescription: '',
-    version: 1,
-    associatedCreditorInstitutions: 0,
-    activationDate: new Date('2023-06-07T16:30:26.384Z'),
-    createdAt: new Date('2023-06-07T16:30:26.384Z'),
-    modifiedAt: new Date('2023-06-07T16:30:26.384Z'),
-    redirectIp: '11.22.44',
-    redirectPath: 'Stazione/path/redirect/prova',
-    redirectPort: 3333,
-    redirectQueryString: 'nessuno',
-    redirectProtocol: RedirectProtocolEnum.HTTPS,
-    brokerCode: '81001870922',
-    timeoutA: 7,
-    timeoutB: 30,
-    timeoutC: 120,
-    targetHost: '33.55.66',
-    targetPort: 4443,
-    targetPath: 'Stazione/path/target/prova',
-    primitiveVersion: 1,
-    wrapperStatus: WrapperStatusEnum.APPROVED,
-    password: genPassword,
+jest.mock('../../../../components/commonFunctions');
+jest.mock("../../../../../hooks/useUserRole");
+
+beforeEach(() => {
+  jest.spyOn(console, 'error').mockImplementation(() => {});
+  jest.spyOn(console, 'warn').mockImplementation(() => {});
+});
+
+afterEach(cleanup);
+
+export const mockedFullStation: StationDetailResource = {
+  associatedCreditorInstitutions: 0,
+  enabled: true,
+  stationCode: '97735020584_01',
+  wrapperStatus: WrapperStatusEnum.APPROVED,
+  brokerCode: '97735020584',
+  ip: 'Valore',
+  ip4Mod: 'Valore',
+  password: genPassword,
+  port: 100,
+  port4Mod: 100,
+  primitiveVersion: 1,
+  protocol: ProtocolEnum.HTTPS,
+  protocol4Mod: Protocol4ModEnum.HTTPS,
+  redirectIp: 'Valore',
+  redirectPath: 'Valore',
+  redirectPort: 8080,
+  redirectProtocol: RedirectProtocolEnum.HTTPS,
+  redirectQueryString: 'Valore',
+  service: 'Valore',
+  service4Mod: 'Valore',
+  targetHost: 'Valore',
+  targetPath: 'Valore',
+  targetPort: 1000,
+  version: 2,
+  pofService: "pofService"
 };
 
-export const mockedFullStationToCheck: StationDetailResource = {
-    stationCode: '81001870922_06',
-    enabled: true,
-    brokerDescription: '',
-    version: 1,
-    associatedCreditorInstitutions: 0,
-    activationDate: new Date('2023-06-07T16:30:26.384Z'),
-    createdAt: new Date('2023-06-07T16:30:26.384Z'),
-    modifiedAt: new Date('2023-06-07T16:30:26.384Z'),
-    redirectIp: '11.22.44',
-    redirectPath: 'Stazione/path/redirect/prova',
-    redirectPort: 3333,
-    redirectQueryString: 'nessuno',
-    redirectProtocol: RedirectProtocolEnum.HTTPS,
-    brokerCode: '81001870922',
-    timeoutA: 7,
-    timeoutB: 30,
-    timeoutC: 120,
-    targetHost: '33.55.66',
-    targetPort: 4443,
-    targetPath: 'Stazione/path/target/prova',
-    primitiveVersion: 1,
-    wrapperStatus: WrapperStatusEnum.TO_CHECK,
-};
+describe('<StationDetails.test />', () => {
+  createMemoryHistory();
 
-describe('<StationDetails />', () => {
-    const history = createMemoryHistory();
-
-    test('render component StationDetails and exit button test', () => {
-        store.dispatch(partiesActions.setPartySelected(ecAdminSignedDirect));
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PAGOPA_OPERATOR,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: true,
-        });
-        render(
-            <Provider store={store}>
-                <MemoryRouter initialEntries={[`/stations/${mockedFullStationToCheck.stationCode}`]}>
-                    <Route path="/stations/:stationId">
-                        <ThemeProvider theme={theme}>
-                            <StationDetails
-                                stationDetail={mockedFullStationApproved}
-                                // @ts-ignore TODO
-                                formatedDate={jest.fn()}
-                                goBack={() => jest.fn()}
-                            />
-                        </ThemeProvider>
-                    </Route>
-                </MemoryRouter>
-            </Provider>
-        );
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PSP_ADMIN,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: false,
-            userIsAdmin: false
-        });
-        if (!useUserRole.useUserRole().userIsPagopaOperator) {
-            const backBtn = screen.getByTestId('exit-btn-test');
-            fireEvent.click(backBtn);
-            expect(history.location.pathname).toBe('/');
-        }
+  test('render component StationDetails', () => {
+    jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
+      userRole: ROLE.PAGOPA_OPERATOR,
+      userIsPspAdmin: false,
+      userIsEcAdmin: false,
+      userIsPspDirectAdmin: false,
+      userIsPagopaOperator: true,
+      userIsAdmin: true
     });
+    render(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={[`/stations/${mockedFullStation.stationCode}`]}>
+          <Route path="/stations/:stationId">
+            <ThemeProvider theme={theme}>
+              <StationDetails
+                stationDetail={mockedFullStation}
+                // @ts-ignore TODO
+                formatedDate={jest.fn()}
+              />
+            </ThemeProvider>
+          </Route>
+        </MemoryRouter>
+      </Provider>
+    );
+    const title = screen.getAllByText('97735020584_01');
+    expect(title[0]).toBeInTheDocument();
+  });
 
-    test('Test edit Button with StationDetails in role operator and status approved', async () => {
-        store.dispatch(partiesActions.setPartySelected(ecAdminSignedDirect));
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PAGOPA_OPERATOR,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: true,
-        });
-        render(
-            <Provider store={store}>
-                <MemoryRouter initialEntries={[`/stations/${mockedFullStationApproved.stationCode}`]}>
-                    <Route path="/stations/:stationId">
-                        <ThemeProvider theme={theme}>
-                            <StationDetails
-                                stationDetail={mockedFullStationApproved}
-                                // @ts-ignore TODO
-                                formatedDate={jest.fn()}
-                                goBack={() => jest.fn()}
-                            />
-                        </ThemeProvider>
-                    </Route>
-                </MemoryRouter>
-            </Provider>
-        );
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PAGOPA_OPERATOR,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: true
-        });
-        if (!useUserRole.useUserRole().userIsPagopaOperator) {
-            const editBtn = await screen.findByTestId('edit-btn-sts-approved');
-            fireEvent.click(editBtn);
-        }
+  test('Test edit Button with StationDetails in role operator and status approved', async () => {
+    jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
+      userRole: ROLE.PAGOPA_OPERATOR,
+      userIsPspAdmin: false,
+      userIsEcAdmin: false,
+      userIsPspDirectAdmin: false,
+      userIsPagopaOperator: true,
+      userIsAdmin: true
     });
+    render(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={[`/stations/${mockedFullStation.stationCode}`]}>
+          <Route path="/stations/:stationId">
+            <ThemeProvider theme={theme}>
+              <StationDetails
+                stationDetail={mockedFullStation}
+                // @ts-ignore TODO
+                formatedDate={jest.fn()}
+              />
+            </ThemeProvider>
+          </Route>
+        </MemoryRouter>
+      </Provider>
+    );
+  });
 
-    test('render component StationDetails', async () => {
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PAGOPA_OPERATOR,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: true,
-        });
-        render(
-            <Provider store={store}>
-                <MemoryRouter initialEntries={[`/stations/${mockedFullStationToCheck.stationCode}`]}>
-                    <Route path="/stations/:stationId">
-                        <ThemeProvider theme={theme}>
-                            <StationDetails
-                                stationDetail={mockedFullStationToCheck}
-                                // @ts-ignore TODO
-                                formatedDate={jest.fn()}
-                                goBack={() => jest.fn()}
-                            />
-                        </ThemeProvider>
-                    </Route>
-                </MemoryRouter>
-            </Provider>
-        );
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PSP_ADMIN,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: false
-        });
-        if (!useUserRole.useUserRole().userIsPagopaOperator) {
-            const editBtn = await screen.findByTestId('edit-btn-sts-approved');
-            fireEvent.click(editBtn);
-        }
+  test('Test show password button', async () => {
+    jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
+      userRole: ROLE.PAGOPA_OPERATOR,
+      userIsPspAdmin: false,
+      userIsEcAdmin: false,
+      userIsPspDirectAdmin: false,
+      userIsPagopaOperator: true,
+      userIsAdmin: true
     });
+    render(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={[`/stations/${mockedFullStation.stationCode}`]}>
+          <Route path="/stations/:stationId">
+            <ThemeProvider theme={theme}>
+              <StationDetails
+                stationDetail={mockedFullStation}
+                // @ts-ignore TODO
+                formatedDate={jest.fn()}
+              />
+            </ThemeProvider>
+          </Route>
+        </MemoryRouter>
+      </Provider>
+    );
 
-    test('render component StationDetails and test ShowHidePassword button', async () => {
-        jest.spyOn(useUserRole, 'useUserRole').mockReturnValue({
-            userRole: ROLE.PAGOPA_OPERATOR,
-            userIsPspAdmin: false,
-            userIsEcAdmin: false,
-            userIsPspDirectAdmin: false,
-            userIsPagopaOperator: true,
-            userIsAdmin: true,
-        });
-        render(
-            <Provider store={store}>
-                <MemoryRouter initialEntries={[`/stations/${mockedFullStationApproved.stationCode}`]}>
-                    <Route path="/stations/:stationId">
-                        <ThemeProvider theme={theme}>
-                            <StationDetails
-                                stationDetail={mockedFullStationApproved}
-                                // @ts-ignore TODO
-                                formatedDate={jest.fn()}
-                                goBack={() => jest.fn()}
-                            />
-                        </ThemeProvider>
-                    </Route>
-                </MemoryRouter>
-            </Provider>
-        );
+    const showPasswordButton = screen.getByTestId('show-pwd-validation-test') as HTMLInputElement;
+
+    expect(screen.getByText('XXXXXXXXXXXXXX')).toBeInTheDocument();
+    expect(screen.queryByText(genPassword)).toBeNull();
+
+    fireEvent.click(showPasswordButton);
+    await waitFor(() => {
+      expect(screen.getByText(genPassword)).toBeInTheDocument();
     });
+  });
 });
